+++ src/plugin-marklogic/main/uk/co/reecedunn/intellij/plugin/marklogic/rewriter/endpoints/RewriterEndpoint.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020-2021 Reece H. Dunn
+ * Copyright (C) 2020-2022 Reece H. Dunn
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -28,7 +28,6 @@ import javax.swing.Icon
 
 class RewriterEndpoint(val endpoint: XmlTag) :
     ItemPresentation,
-    EndpointMethodPresentation,
     DataProvider {
     // region ItemPresentation
 
@@ -41,10 +40,10 @@ class RewriterEndpoint(val endpoint: XmlTag) :
     // endregion
     // region EndpointMethodPresentation
 
-    override val endpointMethod: String?
+    val endpointMethod: String?
         get() = endpoint.ancestor(Rewriter.NAMESPACE, "match-method").firstOrNull()?.getAttributeValue("any-of")
 
-    override val endpointMethodOrder: Int
+    val endpointMethodOrder: Int
         get() = EndpointMethodPresentation.getHttpMethodOrder(endpointMethod?.split("\\s+")?.get(0))
 
     // endregion